refactor(richtext): make quillmark-richtext a leaf crate core depends on
Invert the crate dependency arrow so quillmark-richtext (the RichText corpus
model, canonical serialization, edit deltas, and markdown codecs) is a leaf
that quillmark-core depends on, per the phase-2 topology decision. This is the
groundwork PR-A of issue #831 phase 2.

- Relocate the pure markdown-string normalizer (normalize_markdown + line-ending
  / bidi-strip / HTML-comment-fence helpers) from core into a new
  quillmark-richtext::normalize module. It is a rich-text primitive, not a
  document-engine concern; moving it removes richtext's last dependency on core.
- Relocate MAX_NESTING_DEPTH to quillmark-richtext (its home is the import
  guard); core::error re-exports it so the typst backend's markup converter keeps
  sharing one limit — a document that imports also renders.
- core::normalize keeps the document-level pass (field-name NFC + per-card body
  normalization), now delegating the string step to quillmark_richtext.
- quillmark-richtext drops its quillmark-core dependency and becomes publish=true
  (quillmark-core, published, now depends on it). pulldown is reachable only
  through the codecs, so no-feature pkg/core still tree-shakes it out until a
  body/import path is wired (phase 2 proper).

Also retire the "Quill-Delta semantics" framing on the edit surface: Delta is
retain/insert/delete text splices (CodeMirror ChangeSet / OT), NOT attributed
Quill-Delta ops. Marks and islands rebase THROUGH a delta via map_pos; they are
not op attributes — an attribute map cannot carry overlapping same-kind marks or
distinct identity anchors, the algebra the corpus model keeps. Mark and line
edits are separate op channels (phase 3). Doc-only; no behavior change.

All 712 core+richtext tests pass; full workspace builds; clippy clean.
